pyproject.toml
[tool.poetry]
name = "fastapitemplate"
version = "0.1.0"
description = ""
authors = ["kirill <shkolenko02-03-2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
alembic = "^1.11.3"
loguru = "^0.7.0"
ruff = "^0.3.7"
pre-commit = "^3.7.0"
asyncpg = "^0.29.0"
minio = "^7.2.10"
clickhouse-connect = "^0.8.5"
pillow = "^11.0.0"
torch = "^2.5.0"
python-docx = "^1.1.2"
check-swear = "^0.1.4"
yandex-cloud-ml-sdk = "^0.1.0"
langchain-core = "0.2.29"
numpy = "<2.0"
transformers = "^4.46.0"
open-clip-torch = "^2.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
